% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A function that takes a country name and generates a graph of any country's population over time.}
\usage{
CountryPopulation(countryname)
}
\arguments{
\item{x}{A country name within the data frame}
}
\value{
A graph of \code{countryname} population over time
}
\description{
A function that takes a country name and generates a graph of any country's population over time.
}
\examples{
CountryPopulation('China')
CountryPopulation('India')
}
